+++ example/lib/main.dart
@@ -72,6 +72,7 @@ void main() async {
 
 class MyApp extends StatelessWidget {
   const MyApp({Key? key}) : super(key: key);
+
   // private navigators
 
   @override
@@ -247,6 +248,17 @@ class _MyHomePageState extends State<MyHomePage> with WindowListener {
       onTap: () {
         if (router.location != '/forms/numberbox') {
           router.pushNamed('forms_numberbox');
+        }
+      },
+    ),
+    PaneItem(
+      key: const Key('/forms/passwordbox'),
+      icon: const Icon(FluentIcons.password_field),
+      title: const Text('PasswordBox'),
+      body: const SizedBox.shrink(),
+      onTap: () {
+        if (router.location != '/forms/passwordbox') {
+          router.pushNamed('forms_passwordbox');
         }
       },
     ),
@@ -869,6 +881,15 @@ final router = GoRouter(
           builder: (context, state) => DeferredWidget(
             forms.loadLibrary,
             () => forms.NumberBoxPage(),
+          ),
+        ),
+
+        GoRoute(
+          path: '/forms/passwordbox',
+          name: 'forms_passwordbox',
+          builder: (context, state) => DeferredWidget(
+            forms.loadLibrary,
+            () => forms.PasswordBoxPage(),
           ),
         ),